Removed a lot of cruft, slight reorg.
* Now mostly just contains utilities for training and running jobs in parallel.
  Other stuff can be moved to another repo eventually.

+++ dps/utils/tf.py
@@ -1780,7 +1780,7 @@ class Polynomial(Schedule):
         assert power > 0
 
     def build(self, t):
-        t = tf.minimum(tf.cast(self.decay_steps, tf.int64), t)
+        t = tf.minimum(tf.cast(self.decay_steps, t.dtype), t)
         return (self.start - self.end) * ((1 - t / self.decay_steps) ** self.power) + self.end
